Move volume validation code to its own file
